+++ storybook-static/assets/DAccordion.stories-db7c1f83.js
@@ -1,0 +1,12 @@
+import{D as t}from"./DAccordion-bf179bbb.js";import"./vue.esm-bundler-02a62393.js";import"./_commonjsHelpers-87174ba5.js";const d={title:"Accordion",component:t,argTypes:{}},s=a=>({components:{DAccordion:t},setup(){return{args:a}},template:'<d-accordion v-bind="args" />'}),o=s.bind({});o.args={title:"Personal information verification"};var r,n,e;o.parameters={...o.parameters,docs:{...(r=o.parameters)==null?void 0:r.docs,source:{originalSource:`args => ({
+  components: {
+    DAccordion
+  },
+  setup() {
+    return {
+      args
+    };
+  },
+  template: \`<d-accordion v-bind="args" />\`
+})`,...(e=(n=o.parameters)==null?void 0:n.docs)==null?void 0:e.source}}};const m=["Default"];export{o as Default,m as __namedExportsOrder,d as default};
+//# sourceMappingURL=DAccordion.stories-db7c1f83.js.map